moved FileUtil to package resources.files
