+++ src/main/java/tech/fastj/resources/files/FileUtil.java
@@ -1,4 +1,4 @@
-package tech.fastj.systems.fio;
+package tech.fastj.resources.files;
 
 import java.nio.file.Path;
